Make OpenTelemetry work in production build (#5)
Next.js’s Output File Tracing messes with the built output files.

`require.resolve` is used to force some files to be included in the
built output. Also, instead of requiring the instrumentation script from
within endpoints (which causes it to gets bundled, but in a weird way),
we simply `require.resolve` it and then use `node -r` to load the
instrumentation script.

One particular problems faced is that file tracing does not copy
packages in `packages/*` to the resulting
`.next/standalone/node_modules` directory, so as a hacky workaround, a
postbuild script is added to manually copy otel package files into
`.next/standalone/node_modules` folder.